[feat] changeToReactNativeMaps
