--- backendTest.js
@@ -1,480 +1,0 @@
-// backendTest.js
-export const result = [
-  {
-    id: 1,
-    name: null,
-    points: [
-      {
-        id: 1,
-        lng: 126.9541171831812,
-        lat: 37.5050168231913,
-        pointType: null,
-        routes: [],
-      },
-      {
-        id: 3,
-        lng: 126.9547609133448,
-        lat: 37.5051444908722,
-        pointType: null,
-        routes: [],
-      },
-      {
-        id: 6,
-        lng: 126.95534563491,
-        lat: 37.5048806440909,
-        pointType: null,
-        routes: [],
-      },
-      {
-        id: 7,
-        lng: 126.9552651686396,
-        lat: 37.5046040295285,
-        pointType: null,
-        routes: [],
-      },
-      {
-        id: 8,
-        lng: 126.9563165945734,
-        lat: 37.504293368875,
-        pointType: null,
-        routes: [],
-      },
-      {
-        id: 9,
-        lng: 126.9570086044993,
-        lat: 37.504459338426,
-        pointType: null,
-        routes: [],
-      },
-    ],
-    roads: [
-      {
-        id: 1,
-        startLng: 126.9541171831812,
-        startLat: 37.5050168231913,
-        endLng: 126.9547609133448,
-        endLat: 37.5051444908722,
-        totalRoadDistance: 58.59716995552399,
-        routes: [],
-        street: false,
-        lane: true,
-        steep: false,
-      },
-      {
-        id: 2,
-        startLng: 126.9547609133448,
-        startLat: 37.5051444908722,
-        endLng: 126.95534563491,
-        endLat: 37.5048806440909,
-        totalRoadDistance: 59.40549914608634,
-        routes: [],
-        street: false,
-        lane: true,
-        steep: false,
-      },
-      {
-        id: 5,
-        startLng: 126.95534563491,
-        startLat: 37.5048806440909,
-        endLng: 126.9552651686396,
-        endLat: 37.5046040295285,
-        totalRoadDistance: 31.60187855476608,
-        routes: [],
-        street: false,
-        lane: true,
-        steep: false,
-      },
-      {
-        id: 9,
-        startLng: 126.9552651686396,
-        startLat: 37.5046040295285,
-        endLng: 126.9563165945734,
-        endLat: 37.504293368875,
-        totalRoadDistance: 99.08292364431279,
-        routes: [],
-        street: false,
-        lane: true,
-        steep: false,
-      },
-      {
-        id: 10,
-        startLng: 126.9563165945734,
-        startLat: 37.504293368875,
-        endLng: 126.9570086044993,
-        endLat: 37.504459338426,
-        totalRoadDistance: 63.843525172847535,
-        routes: [],
-        street: false,
-        lane: true,
-        steep: false,
-      },
-    ],
-    totalRouteDistance: 312.53,
-    totalRouteTime: 224,
-    routeType: null,
-    user: null,
-  },
-  {
-    id: 2,
-    name: null,
-    points: [
-      {
-        id: 1,
-        lng: 126.9541171831812,
-        lat: 37.5050168231913,
-        pointType: null,
-        routes: [],
-      },
-      {
-        id: 3,
-        lng: 126.9547609133448,
-        lat: 37.5051444908722,
-        pointType: null,
-        routes: [],
-      },
-      {
-        id: 4,
-        lng: 126.9553724570002,
-        lat: 37.5051402352864,
-        pointType: null,
-        routes: [],
-      },
-      {
-        id: 5,
-        lng: 126.9559679074015,
-        lat: 37.5049061776911,
-        pointType: null,
-        routes: [],
-      },
-      {
-        id: 9,
-        lng: 126.9570086044993,
-        lat: 37.504459338426,
-        pointType: null,
-        routes: [],
-      },
-    ],
-    roads: [
-      {
-        id: 1,
-        startLng: 126.9541171831812,
-        startLat: 37.5050168231913,
-        endLng: 126.9547609133448,
-        endLat: 37.5051444908722,
-        totalRoadDistance: 58.59716995552399,
-        routes: [],
-        street: false,
-        lane: true,
-        steep: false,
-      },
-      {
-        id: 6,
-        startLng: 126.9547609133448,
-        startLat: 37.5051444908722,
-        endLng: 126.9553724570002,
-        endLat: 37.5051402352864,
-        totalRoadDistance: 54.00725780034931,
-        routes: [],
-        street: false,
-        lane: true,
-        steep: false,
-      },
-      {
-        id: 7,
-        startLng: 126.9553724570002,
-        startLat: 37.5051402352864,
-        endLng: 126.9559679074015,
-        endLat: 37.5049061776911,
-        totalRoadDistance: 58.685236255699195,
-        routes: [],
-        street: false,
-        lane: true,
-        steep: false,
-      },
-      {
-        id: 8,
-        startLng: 126.9559679074015,
-        startLat: 37.5049061776911,
-        endLng: 126.9570086044993,
-        endLat: 37.504459338426,
-        totalRoadDistance: 104.50179628958686,
-        routes: [],
-        street: false,
-        lane: true,
-        steep: false,
-      },
-    ],
-    totalRouteDistance: 275.79,
-    totalRouteTime: 197,
-    routeType: null,
-    user: null,
-  },
-  {
-    id: 3,
-    name: null,
-    points: [
-      {
-        id: 1,
-        lng: 126.9541171831812,
-        lat: 37.5050168231913,
-        pointType: null,
-        routes: [],
-      },
-      {
-        id: 2,
-        lng: 126.9545624298777,
-        lat: 37.5048295768642,
-        pointType: null,
-        routes: [],
-      },
-      {
-        id: 7,
-        lng: 126.9552651686396,
-        lat: 37.5046040295285,
-        pointType: null,
-        routes: [],
-      },
-      {
-        id: 6,
-        lng: 126.95534563491,
-        lat: 37.5048806440909,
-        pointType: null,
-        routes: [],
-      },
-      {
-        id: 7,
-        lng: 126.9552651686396,
-        lat: 37.5046040295285,
-        pointType: null,
-        routes: [],
-      },
-      {
-        id: 3,
-        lng: 126.9547609133448,
-        lat: 37.5051444908722,
-        pointType: null,
-        routes: [],
-      },
-      {
-        id: 6,
-        lng: 126.95534563491,
-        lat: 37.5048806440909,
-        pointType: null,
-        routes: [],
-      },
-      {
-        id: 3,
-        lng: 126.9547609133448,
-        lat: 37.5051444908722,
-        pointType: null,
-        routes: [],
-      },
-      {
-        id: 4,
-        lng: 126.9553724570002,
-        lat: 37.5051402352864,
-        pointType: null,
-        routes: [],
-      },
-      {
-        id: 5,
-        lng: 126.9559679074015,
-        lat: 37.5049061776911,
-        pointType: null,
-        routes: [],
-      },
-      {
-        id: 9,
-        lng: 126.9570086044993,
-        lat: 37.504459338426,
-        pointType: null,
-        routes: [],
-      },
-    ],
-    roads: [
-      {
-        id: 3,
-        startLng: 126.9541171831812,
-        startLat: 37.5050168231913,
-        endLng: 126.9545624298777,
-        endLat: 37.5048295768642,
-        totalRoadDistance: 44.50299087275961,
-        routes: [],
-        street: false,
-        lane: true,
-        steep: false,
-      },
-      {
-        id: 4,
-        startLng: 126.9545624298777,
-        startLat: 37.5048295768642,
-        endLng: 126.9552651686396,
-        endLat: 37.5046040295285,
-        totalRoadDistance: 66.94560569617953,
-        routes: [],
-        street: false,
-        lane: true,
-        steep: false,
-      },
-      {
-        id: 5,
-        startLng: 126.95534563491,
-        startLat: 37.5048806440909,
-        endLng: 126.9552651686396,
-        endLat: 37.5046040295285,
-        totalRoadDistance: 31.60187855476608,
-        routes: [],
-        street: false,
-        lane: true,
-        steep: false,
-      },
-      {
-        id: 2,
-        startLng: 126.9547609133448,
-        startLat: 37.5051444908722,
-        endLng: 126.95534563491,
-        endLat: 37.5048806440909,
-        totalRoadDistance: 59.40549914608634,
-        routes: [],
-        street: false,
-        lane: true,
-        steep: false,
-      },
-      {
-        id: 6,
-        startLng: 126.9547609133448,
-        startLat: 37.5051444908722,
-        endLng: 126.9553724570002,
-        endLat: 37.5051402352864,
-        totalRoadDistance: 54.00725780034931,
-        routes: [],
-        street: false,
-        lane: true,
-        steep: false,
-      },
-      {
-        id: 7,
-        startLng: 126.9553724570002,
-        startLat: 37.5051402352864,
-        endLng: 126.9559679074015,
-        endLat: 37.5049061776911,
-        totalRoadDistance: 58.685236255699195,
-        routes: [],
-        street: false,
-        lane: true,
-        steep: false,
-      },
-      {
-        id: 8,
-        startLng: 126.9559679074015,
-        startLat: 37.5049061776911,
-        endLng: 126.9570086044993,
-        endLat: 37.504459338426,
-        totalRoadDistance: 104.50179628958686,
-        routes: [],
-        street: false,
-        lane: true,
-        steep: false,
-      },
-    ],
-    totalRouteDistance: 419.65,
-    totalRouteTime: 300,
-    routeType: null,
-    user: null,
-  },
-  {
-    id: 4,
-    name: null,
-    points: [
-      {
-        id: 1,
-        lng: 126.9541171831812,
-        lat: 37.5050168231913,
-        pointType: null,
-        routes: [],
-      },
-      {
-        id: 2,
-        lng: 126.9545624298777,
-        lat: 37.5048295768642,
-        pointType: null,
-        routes: [],
-      },
-      {
-        id: 7,
-        lng: 126.9552651686396,
-        lat: 37.5046040295285,
-        pointType: null,
-        routes: [],
-      },
-      {
-        id: 8,
-        lng: 126.9563165945734,
-        lat: 37.504293368875,
-        pointType: null,
-        routes: [],
-      },
-      {
-        id: 9,
-        lng: 126.9570086044993,
-        lat: 37.504459338426,
-        pointType: null,
-        routes: [],
-      },
-    ],
-    roads: [
-      {
-        id: 3,
-        startLng: 126.9541171831812,
-        startLat: 37.5050168231913,
-        endLng: 126.9545624298777,
-        endLat: 37.5048295768642,
-        totalRoadDistance: 44.50299087275961,
-        routes: [],
-        street: false,
-        lane: true,
-        steep: false,
-      },
-      {
-        id: 4,
-        startLng: 126.9545624298777,
-        startLat: 37.5048295768642,
-        endLng: 126.9552651686396,
-        endLat: 37.5046040295285,
-        totalRoadDistance: 66.94560569617953,
-        routes: [],
-        street: false,
-        lane: true,
-        steep: false,
-      },
-      {
-        id: 9,
-        startLng: 126.9552651686396,
-        startLat: 37.5046040295285,
-        endLng: 126.9563165945734,
-        endLat: 37.504293368875,
-        totalRoadDistance: 99.08292364431279,
-        routes: [],
-        street: false,
-        lane: true,
-        steep: false,
-      },
-      {
-        id: 10,
-        startLng: 126.9563165945734,
-        startLat: 37.504293368875,
-        endLng: 126.9570086044993,
-        endLat: 37.504459338426,
-        totalRoadDistance: 63.843525172847535,
-        routes: [],
-        street: false,
-        lane: true,
-        steep: false,
-      },
-    ],
-    totalRouteDistance: 274.38,
-    totalRouteTime: 196,
-    routeType: null,
-    user: null,
-  },
-];
